Move types up to Windows.Core
